+++ pydantic_scim2/__init__.py
@@ -2,6 +2,7 @@ from .rfc7643.enterprise_user import EnterpriseUser
 from .rfc7643.enterprise_user import Manager
 from .rfc7643.group import Group
 from .rfc7643.group import GroupMember
+from .rfc7643.resource import AnyResource
 from .rfc7643.resource import Meta
 from .rfc7643.resource import Resource
 from .rfc7643.resource_type import ResourceType
@@ -52,6 +53,7 @@ from .rfc7644.search_request import SortOrder
 
 __all__ = [
     "Address",
+    "AnyResource",
     "Attribute",
     "AuthenticationScheme",
     "Bulk",